# mmcsd_sdio: Arm the write complete detection