@davids5 2020-09-10
@btashton On the initial push I had

![image](https://user-images.githubusercontent.com/1945821/92718899-47a4c700-f317-11ea-91a0-ad51e7b6f917.png)

 any idea what caused this and how I can avoid it?